Introduce status to talks
This commit introduces a few possible status values for talk models.
Existing talks with tracks are migrated to status PUBLISHED all
other talks have their status set to DRAFT. The idea is that talks go
through the following status transitions in the future:

-> DRAFT               when a talk is submitted
DRAFT -> SELECTED      when the committee selects a talk
SELECTED -> CONFIRMED  when a speaker accepts/confirms the committee
                       vote
CONFIRMED -> PUBLISHED when the talk is published for voting by the
                       attendees or directly to the session grid

Talks that do not have a PUBLISHED status may get deleted after an event
if the potential speakers delete their accounts. Talks that are thought
as fallback stay in the CONFIRMED status.

New talk.models.Talk instances are created with status DRAFT.

talk.models.Talk.publish sets the status to PUBLISHED now.

This is related to #125 and #37.